Application layer Is done

+++ LaptopParhizgar/src/Domain/Orders/OrderDetail.cs
@@ -8,13 +8,11 @@ public class OrderDetail : BaseEntity
     public long Price { get; private set; }
     public int Count { get; private set; }
 
-    public OrderDetail(int orderId, int productId, int price, int count)
+    public OrderDetail(long orderId, long productId, long price, int count)
     {
         OrderId = orderId;
         ProductId = productId;
         Price = price;
         Count = count;
     }
-
-
 }
